src/requirements.txt
fastapi==0.120.1
uvicorn[standard]==0.38.0
python-multipart== 0.0.20
python-dotenv==1.2.1
pydantic-settings==2.11.0
aiofiles==25.1.0
langchain==0.3.27
langchain-community==0.3.31
PyMuPDF==1.26.5
motor==3.7.1
openai==2.6.1
cohere==5.20.0
qdrant-client==1.15.1
SQLAlchemy==2.0.44
asyncpg==0.30.0
alembic==1.15.1
psycopg2==2.9.11
pgvector==0.4.1
nltk==3.9.2
prometheus-client==0.23.1
starlette-exporter==0.23.0
fastapi-health==0.4.0